+++ test/runtests.jl
@@ -3,7 +3,9 @@ using Base.Test
 
 include("peakmeas.jl")
 include("baseline.jl")
+include("baseline2.jl")
 include("tlcorrection.jl")
 include("polynoms.jl")
 include("peaks.jl")
+include("smooth.jl")